Issue #31 Add Exponential Backoff
This is an initial Commit to provide Exponential Backoff Functionality.
It was quite a fight to understand the ConfigurableHttpClient of Google,
which makes some things quite complicated (at least for me).

This is a Work in Progress Commit, which still needs some sane default
values for the BackOff Algorithm and make it configurable from the
Outside.